feat: optimize table rendering
